[nixos] fix bootloader entries cleanup

+++ nixos/configuration.nix
@@ -122,7 +122,7 @@
 
           cd /boot/loader/entries
 
-          rm $files
+          [[ ! -z $files ]] && rm $files
         '';
       };
     };
